rename world-based modules
