+++ src/game/game_objects.rs
@@ -2,7 +2,7 @@ use crate::entity::genetics::{DnaType, GeneLibrary, GENOME_LEN};
 use crate::entity::object::Object;
 use crate::entity::player::PLAYER;
 use crate::game::position::Position;
-use crate::game::world::Tile;
+use crate::game::world_gen::Tile;
 use crate::game::{WORLD_HEIGHT, WORLD_WIDTH};
 use crate::rand::Rng;
 use crate::util::rng::GameRng;
