+++ src/test/ai.rs
@@ -2,7 +2,7 @@ use crate::entity::control::Controller;
 use crate::entity::genetics::{Actuators, Dna, Processors, Sensors};
 use crate::game::game_state::GameState;
 use crate::game::innit_env;
-use crate::world::Tile;
+use crate::world_gen::Tile;
 use crate::{entity::action::hereditary::ActMove, game::game_objects::GameObjects};
 
 #[test]
