+++ src/world_gen/mod.rs
@@ -1,17 +1,14 @@
 //! The world generation module contains the trait that all world generators have to implement to
 //! be changeably used to create the game environments.
 
-pub mod world_gen_organic;
-pub mod world_gen_rogue;
+pub mod ca;
+pub mod rogue;
 
-use crate::entity::ai::AiTile;
-use crate::entity::control::Controller;
-use crate::entity::object::Object;
+use crate::entity::{ai, control, Object};
+use crate::game;
 use crate::game::game_objects::GameObjects;
 use crate::game::game_state::GameState;
-use crate::game::RunState;
-use crate::raws::object_template::ObjectTemplate;
-use crate::raws::spawn::Spawn;
+use crate::raws;
 use serde::{Deserialize, Serialize};
 
 /// The world generation trait only requests to implement a method that
@@ -24,9 +21,9 @@ pub trait WorldGen {
         &mut self,
         state: &mut GameState,
         objects: &mut GameObjects,
-        spawns: &[Spawn],
-        object_templates: &[ObjectTemplate],
-    ) -> RunState;
+        spawns: &[raws::spawn::Spawn],
+        object_templates: &[raws::object_template::ObjectTemplate],
+    ) -> game::RunState;
 
     /// Returns the position of where the player was placed.
     fn get_player_start_pos(&self) -> (i32, i32);
@@ -57,7 +54,7 @@ impl Tile {
             .visualize("wall tile", '◘', (255, 255, 255, 255))
             .physical(true, true, is_visible)
             .tile_explored(is_visible)
-            .control(Controller::Npc(Box::new(AiTile)))
+            .control(control::Controller::Npc(Box::new(ai::AiTile)))
     }
 }
